try FreeYOLOvx with AlignOTA

+++ models/yolo_free_vx/loss.py
@@ -82,7 +82,7 @@ class Criterion(object):
                 gt_classes = F.one_hot(gt_matched_classes.long(), self.num_classes)
                 gt_classes = gt_classes * pred_ious_this_matching.unsqueeze(-1)
                 cls_target = cls_preds.new_zeros((num_anchors, self.num_classes))
-                cls_target[fg_mask] = gt_classes
+                cls_target[fg_mask] = gt_classes.float()
                 # box target
                 box_target = tgt_bboxes[matched_gt_inds]
